Moves diagnostics related feature components from the general components folder to @/features/diagnostics/
