+++ features/diagnostics/PatientTestsSection/index.ts
@@ -1,0 +1,1 @@
+export { PatientTestsSection } from './PatientTestsSection'
