+++ features/diagnostics/TestChart/index.ts
@@ -1,0 +1,1 @@
+export { TestChart } from './TestChart'
