+++ features/diagnostics/TestDetails/index.ts
@@ -1,0 +1,1 @@
+export { TestDetails } from './TestDetails'
