+++ features/diagnostics/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Diagnostics feature exports
+ * Centralized exports for all diagnostic-related components
+ */
+
+export { PatientTestsSection } from './PatientTestsSection'
+export { TestChart } from './TestChart'
+export { TestDetails } from './TestDetails'
